+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improved error handling
 - Improved MMRL app support
 - Added `repo` subcommand
+- Added support to install local modules
 - Removed download progress bar due to deadlock in MMRL
 - Other improvements
 - Added ModConf